Using common project for JWT token instead of new module
