+++ content-api/content-service/conf/application.conf
@@ -545,10 +545,10 @@ learning.service_provider=["youtube"]
 stream.mime.type=video/mp4
 compositesearch.index.name="compositesearch"
 
-hierarchy.keyspace.name=hierarchy_store
+hierarchy.keyspace.name=dev_hierarchy_store
 content.hierarchy.table=content_hierarchy
 framework.hierarchy.table=framework_hierarchy
-objectcategorydefinition.keyspace=category_store
+objectcategorydefinition.keyspace=dev_category_store
 
 # Kafka topic for definition update event.
 kafka.topic.system.command="dev.system.command"
