+++ content-api/hierarchy-manager/pom.xml
@@ -14,11 +14,6 @@
     <dependencies>
         <dependency>
             <groupId>org.sunbird</groupId>
-            <artifactId>auth-verifier</artifactId>
-            <version>1.0-SNAPSHOT</version>
-        </dependency>
-        <dependency>
-            <groupId>org.sunbird</groupId>
             <artifactId>graph-engine_2.11</artifactId>
             <version>1.0-SNAPSHOT</version>
             <type>jar</type>
