+++ pom.xml
@@ -16,9 +16,6 @@
 		<scala.version>2.11.12</scala.version>
 		<scalatest.version>3.0.8</scalatest.version>
 		<fasterxml.jackson.version>2.9.8</fasterxml.jackson.version>
-		<mockito.core.version>2.22.0</mockito.core.version>
-		<powermock.api.mockito2.version>2.0.7</powermock.api.mockito2.version>
-		<powermock.module.junit4.version>2.0.0-beta.5</powermock.module.junit4.version>
 	</properties>
 	<modules>
 		<module>platform-core</module>
